backend/requirements.txt
fastapi[standard]
uvicorn[standard]
websockets>=13.0
python-dotenv
google-genai>=1.0
httpx
starlette